+++ spec/chewy/index/actions_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 describe Chewy::Index::Actions do
   include ClassHelpers
-  before { Chewy::Index.client.indices.delete }
+  before { Chewy.client.indices.delete index: '*' }
 
   before { stub_index :dummies }