+++ spec/chewy/rspec/update_index_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 describe :update_index do
   include ClassHelpers
-  before { Chewy::Index.client.indices.delete }
+  before { Chewy.client.indices.delete index: '*' }
 
   before do
     stub_index(:dummies) do