+++ spec/chewy/type/import_spec.rb
@@ -3,7 +3,7 @@ require 'spec_helper'
 describe Chewy::Type::Import do
   include ClassHelpers
 
-  before { Chewy.client.indices.delete }
+  before { Chewy.client.indices.delete index: '*' }
 
   before do
     stub_model(:city)